fdt_locales_lst {bigD}	R Documentation

A list of all supported locales

Description

The fdt_locales_lst object is a list of all supported locales. This is useful when used with the fdt() function as the list can be auto-completed with a locale identifier and this generates valid input for the locale argument.

Usage

fdt_locales_lst

Format

An object of class list of length 574.

Value

A list where each element corresponds to a supported locale ID.

Examples

The fdt_locales_lst object can be incredibly useful when choosing one of supported locales. You can avoid typing errors and every element of the list is meant to work. In this example, we'll use the "da" locale through use of the list.

fdt(
  input = "2018-07-04 22:05",
  format = "yy-MMMM-d",
  locale = fdt_locales_lst$da
)

#> [1] "18-juli-4"

[Package bigD version 0.3.1 Index]