fdt_locales_vec {bigD}	R Documentation

Get a vector of all supported locales

Description

The fdt_locales_vec() function produces a vector of all supported locale IDs in the bigD package.

Usage

fdt_locales_vec()

Value

A character vector of supported locale IDs.

Examples

# Let's get all the `ar` locales that exist
# in the vector produced by `fdt_locales_vec()`
grep("^ar", fdt_locales_vec(), value = TRUE)

# Let's get all the locales that pertain to the
# `CH` territory in the vector produced by
# `fdt_locales_vec()`
grep("CH", fdt_locales_vec(), value = TRUE)

[Package bigD version 0.3.1 Index]