flex_d_vec {bigD}	R Documentation

Get a vector of all flexible date types

Description

The flex_d_vec() function produces a vector of all supported flexible date types in the bigD package. These types are essentially identifiers for classes of cross-locale date formatting, so, none of these should be used directly in the format argument of the fdt() function (use the flex_d_lst object for that).

Usage

flex_d_vec()

Value

A character vector of supported flexible date types.

[Package bigD version 0.3.1 Index]