names_months {bigD}	R Documentation

Get a vector of all the short month names

Description

The names_months() function produces a vector of all short month names used by the bigD package.

Usage

names_months()

Value

A character vector of short month names.

Examples

# Let's get all the short month names with
# the `names_months()` function
names_months()

[Package bigD version 0.3.1 Index]