names_wkdays {bigD}	R Documentation

Get a vector of all the short weekday names

Description

The names_wkdays() function produces a vector of all short weekday names used by the bigD package.

Usage

names_wkdays()

Value

A character vector of short weekday names.

Examples

# Let's get all the short weekday names with
# the `names_wkdays()` function
names_wkdays()

[Package bigD version 0.3.1 Index]