Get a vector of all supported locales

The fdt_locales_vec() function produces a vector of all supported locale IDs in the bigD package.

Usage

fdt_locales_vec()

Value

A character vector of supported locale IDs.

Examples

# Let's get all the `ar` locales that exist
# in the vector produced by `fdt_locales_vec()`
grep("^ar", fdt_locales_vec(), value = TRUE)
#>  [1] "ar"    "ar_AE" "ar_BH" "ar_DJ" "ar_DZ" "ar_EG" "ar_EH" "ar_ER" "ar_IL"
#> [10] "ar_IQ" "ar_JO" "ar_KM" "ar_KW" "ar_LB" "ar_LY" "ar_MA" "ar_MR" "ar_OM"
#> [19] "ar_PS" "ar_QA" "ar_SA" "ar_SD" "ar_SO" "ar_SS" "ar_SY" "ar_TD" "ar_TN"
#> [28] "ar_YE"

# Let's get all the locales that pertain to the
# `CH` territory in the vector produced by
# `fdt_locales_vec()`
grep("CH", fdt_locales_vec(), value = TRUE)
#> [1] "de_CH" "en_CH" "fr_CH" "it_CH" "pt_CH"