Get a vector of all 12-hour flexible time types

The flex_t12_vec() function produces a vector of all supported flexible 12-hour time types in the bigD package. These types are essentially identifiers for classes of cross-locale time formatting, so, none of these should be used directly in the format argument of the fdt() function (use the flex_t12_lst object for that).

Usage

flex_t12_vec()

Value

A character vector of supported 12-hour flexible time types.